requirements.txt
alabaster==0.7.13
altair==4.2.2
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
attrs==23.2.0
avro==1.11.0
Babel==2.14.0
beautifulsoup4==4.12.2
boto3==1.34.11
botocore==1.34.11
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confluent-kafka==2.1.1
cryptography==41.0.7
dataclasses==0.6
decorator==5.1.1
deprecation==2.1.0
docutils==0.20.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastavro==1.8.2
fastjsonschema==2.19.1
fsspec==2023.12.2
furl==2.1.3
furo==2023.9.10
future==0.18.3
gotrue==2.1.0
great-expectations==0.14.13
h11==0.14.0
hopsworks==3.4.3
hsfs==3.4.5
hsml==3.4.5
httpcore==0.17.3
httpx==0.24.1
idna==3.6
imagesize==1.4.1
importlib-metadata==7.0.1
ipython==8.18.1
javaobj-py3==0.4.3
jedi==0.19.1
Jinja2==3.0.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_core==5.5.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
mock==5.1.0
nbformat==5.9.2
numpy==1.26.2
orderedmultidict==1.0.1
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
polars==0.20.2
postgrest==0.13.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.1
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyHopsHive==0.6.4.1.dev0
pyhumps==1.6.1
pyjks==20.0.0
PyMySQL==1.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
realtime==1.0.2
referencing==0.32.0
requests==2.31.0
rpds-py==0.16.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
sphinx-basic-ng==1.0.0b2
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.24
stack-data==0.6.3
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.0
supafunc==0.3.1
termcolor==2.4.0
thrift==0.16.0
toolz==0.12.0
tqdm==4.66.1
traitlets==5.14.0
twofish==0.3.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
wcwidth==0.2.12
websockets==11.0.3
zipp==3.17.0
